# Subarray GetSuccessors