--- a/src/stinger_core.jl
+++ b/src/stinger_core.jl
@@ -157,14 +157,8 @@ function outdegree(s::Stinger, src::Int64)
 end
 
 "Return a `Vector` of indices representing the successors of the source"
-function getsuccessors(s::Stinger, src::Int64)
-    #Ideally, allocating a buffer would be preffered.
-    #This makes 2 calls to outdegree - one from here and one from inside C
+function getsuccessors(s::Stinger, src::Int64, storage::Array{Int64})
     outdeg = outdegree(s, src)
-    vertexneighbors = zeros(Int64, outdeg)
-    if outdeg==0
-        return vertexneighbors
-    end
     ccall(
         dlsym(stinger_core_lib, "stinger_gather_successors"),
         Void,
@@ -182,14 +176,14 @@ function getsuccessors(s::Stinger, src::Int64)
         s,
         src,
         pointer_from_objref(outdeg),
-        vertexneighbors,
+        storage,
         C_NULL,
         C_NULL,
         C_NULL,
         C_NULL,
         typemax(Int64)
     )
-    vertexneighbors
+    return view(storage, 1:outdeg)
 end
 
 "Return the weight of the edge. If it doesn't exist return 0."