--- a/src/stinger_core.jl
+++ b/src/stinger_core.jl
@@ -158,14 +158,14 @@ end
 
 "Return a `Vector` of indices representing the successors of the source"
 function getsuccessors(s::Stinger, src::Int64, storage::Array{Int64})
-    outdeg = outdegree(s, src)
+    outdeg = Csize_t[0]
     ccall(
         dlsym(stinger_core_lib, "stinger_gather_successors"),
         Void,
         (
             Ptr{Void}, #Stinger instance
             Int64, #Source vertex
-            Ptr{Int64}, #Outdegree variable
+            Ptr{Csize_t}, #Outdegree variable
             Ptr{Int64}, #Output buffer
             Ptr{Int64}, #weight
             Ptr{Int64}, #timefirst
@@ -175,7 +175,7 @@ function getsuccessors(s::Stinger, src::Int64, storage::Array{Int64})
         ),
         s,
         src,
-        pointer_from_objref(outdeg),
+        outdeg,
         storage,
         C_NULL,
         C_NULL,
@@ -183,7 +183,7 @@ function getsuccessors(s::Stinger, src::Int64, storage::Array{Int64})
         C_NULL,
         typemax(Int64)
     )
-    return view(storage, 1:outdeg)
+    return view(storage, 1:outdeg[1])
 end
 
 "Return the weight of the edge. If it doesn't exist return 0."