Update `getsuccessors` to return a `SubArray` of a pre-allocated buffer that is passed in.