@rohitvarkey 2016-12-07
Original implementation:
```julia
julia> bench(10, 16)
INFO: Running BFS benchmark
BenchmarkTools.Trial:
  samples:          6
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  140.65 mb
  allocs estimate:  623116
  minimum time:     1.05 s (0.79% GC)
  median time:      1.10 s (0.87% GC)
  mean time:        1.11 s (1.01% GC)
  maximum time:     1.19 s (0.74% GC)

julia> bench(11, 16)
INFO: Running BFS benchmark
BenchmarkTools.Trial:
  samples:          4
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  270.13 mb
  allocs estimate:  1130824
  minimum time:     2.01 s (0.90% GC)
  median time:      2.08 s (0.94% GC)
  mean time:        2.11 s (0.93% GC)
  maximum time:     2.25 s (0.85% GC)```
```
With Subarrays
```julia
julia> bench(10, 16)
INFO: Running BFS benchmark
BenchmarkTools.Trial:
  samples:          8
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  109.29 mb
  allocs estimate:  1248549
  minimum time:     784.75 ms (1.56% GC)
  median time:      799.07 ms (1.58% GC)
  mean time:        811.67 ms (1.53% GC)
  maximum time:     915.40 ms (1.47% GC)

julia> bench(11, 16)
INFO: Running BFS benchmark
BenchmarkTools.Trial:
  samples:          5
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  202.77 mb
  allocs estimate:  2268367
  minimum time:     1.54 s (1.26% GC)
  median time:      1.57 s (1.28% GC)
  mean time:        1.58 s (1.31% GC)
  maximum time:     1.61 s (1.36% GC)
```